.local/bin/pyscripts/requirements.txt
certifi==2024.12.14
charset-normalizer==3.4.1
fire==0.7.0
idna==3.10
PyYAML==6.0.2
requests==2.32.3
termcolor==2.5.0
urllib3==2.3.0